+++ server/src/modules/Owe/OwesResolver.ts
@@ -3,6 +3,7 @@ import { Owe, OweState } from "../../models/Owe";
 import { firestore } from "../../db/firebase";
 import { Timestamp, DocumentReference } from "@google-cloud/firestore"
 import { getPermalinkFromOwe } from "../../utils/helpers";
+import { mapOweSnapshot } from "./oweResolver/oweSnapshotMap";
 
 @Resolver()
 export class OwesResolver {
@@ -11,21 +12,7 @@ export class OwesResolver {
     async getOwes() {
         const owesSnapshot = await firestore.collectionGroup('owes').get()
         const owes: Array<Owe> = await Promise.all(owesSnapshot.docs.map(async (oweF) => {
-            const oweFData = oweF.data()
-            const oweFCreated: Timestamp = oweFData.created
-            const issedToRef: DocumentReference = oweFData.issuedToRef
-            const permalink = await getPermalinkFromOwe(oweF)
-            const owe: Owe = {
-                id: oweF.id,
-                documenmentRef: oweF.ref,
-                title: oweFData.title,
-                amount: oweFData.amount,
-                state: oweFData.state ?? OweState.CREATED,
-                issuedByID: oweF.ref.parent.parent!.id,
-                issuedToID: issedToRef.id,
-                created: oweFCreated.toDate(),
-                permalink: permalink
-            }
+            const owe: Owe = await mapOweSnapshot(oweF);
             return owe
         }))
         return owes
